Provide option for HTTPS (#10)
* Add click parameters for SSL cert paths

Added parameters for the paths to the SSL

* certificate,
* private key, and
* certificate chain,

then passed these parameters to `start`. If they're present, HTTPS is
used; if not present, HTTP is used.

The `ssl_chain` parameter may not be needed, according to the CherryPy
docs: https://docs.cherrypy.org/en/latest/deploy.html#ssl-support

* Remove spaces before operator

To satisfy flake8.

(Not my fault, man! It came this way from cherrypy.)

* add news fragment

Co-authored-by: mcflugen <mcflugen@gmail.com>